refactor: move lake and doctor dispatch execution out of CLI modules
