+++ src/commands/lake.rs
@@ -2,9 +2,7 @@
 //!
 //! Subcommands: create, list.
 
-use anyhow::{bail, Result};
-use patina::paths;
-use std::path::PathBuf;
+use anyhow::Result;
 
 /// Lake CLI subcommands
 #[derive(Debug, Clone, clap::Subcommand, serde::Serialize, serde::Deserialize)]
@@ -50,96 +48,18 @@ pub fn execute_cli(command: Option<LakeCommands>) -> Result<()> {
     Ok(())
 }
 
+#[allow(dead_code)]
 pub(crate) fn execute_value(command: LakeCommands) -> Result<serde_json::Value> {
-    match command {
-        LakeCommands::Create { name } => {
-            create(&name)?;
-            Ok(serde_json::json!({
-                "child": "lake-manager",
-                "text": format!("Lake '{}' created", name),
-                "data": {"name": name}
-            }))
-        }
-        LakeCommands::List => {
-            let dir = lakes_dir();
-            let mut lakes = Vec::new();
-            if dir.exists() {
-                let mut entries: Vec<_> = std::fs::read_dir(&dir)?
-                    .filter_map(|e| e.ok())
-                    .filter(|e| e.path().join("lake.toml").exists())
-                    .collect();
-                entries.sort_by_key(|e| e.file_name());
-                for entry in entries {
-                    let lake_toml = entry.path().join("lake.toml");
-                    let content = std::fs::read_to_string(&lake_toml).unwrap_or_default();
-                    let name = content
-                        .lines()
-                        .find(|l| l.starts_with("name"))
-                        .and_then(|l| l.split('=').nth(1))
-                        .map(|v| v.trim().trim_matches('"'))
-                        .unwrap_or("?")
-                        .to_string();
-                    let created = content
-                        .lines()
-                        .find(|l| l.starts_with("created_at"))
-                        .and_then(|l| l.split('=').nth(1))
-                        .map(|v| v.trim().trim_matches('"'))
-                        .unwrap_or("?")
-                        .to_string();
-                    lakes.push(serde_json::json!({
-                        "name": name,
-                        "created_at": created,
-                        "path": entry.path().display().to_string()
-                    }));
-                }
-            }
-            Ok(serde_json::json!({
-                "child": "lake-manager",
-                "data": {"lakes": lakes}
-            }))
-        }
-    }
-}
-
-/// Lakes directory: ~/.patina/lakes/
-fn lakes_dir() -> PathBuf {
-    paths::lakes::lakes_dir()
-}
-
-/// Create a new data lake.
-fn create(name: &str) -> Result<()> {
-    validate_lake_name(name)?;
-
-    let lake_dir = lakes_dir().join(name);
-
-    if lake_dir.exists() {
-        bail!("lake '{}' already exists at {}", name, lake_dir.display());
-    }
-
-    std::fs::create_dir_all(&lake_dir)?;
-
-    let now = chrono::Utc::now().to_rfc3339();
-    let config = format!("name = \"{}\"\ncreated_at = \"{}\"", name, now);
-    std::fs::write(lake_dir.join("lake.toml"), config)?;
-
-    eprintln!("Lake \"{}\" created at {}", name, lake_dir.display());
-    Ok(())
-}
-
-fn validate_lake_name(name: &str) -> Result<()> {
-    if name.is_empty()
-        || !name
-            .chars()
-            .all(|c| c.is_alphanumeric() || c == '-' || c == '_')
-    {
-        bail!("lake name must be non-empty and contain only alphanumeric characters, hyphens, or underscores");
-    }
-    Ok(())
+    let runtime_command = match command {
+        LakeCommands::Create { name } => patina::mother::lake_runtime::LakeCommand::Create { name },
+        LakeCommands::List => patina::mother::lake_runtime::LakeCommand::List,
+    };
+    patina::mother::lake_runtime::execute_value(runtime_command)
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use patina::paths;
 
     #[test]
     fn create_and_list_lake() {
@@ -160,20 +80,17 @@ mod tests {
 
     #[test]
     fn invalid_lake_names() {
-        assert!(create("").is_err());
-        assert!(create("has spaces").is_err());
-        assert!(create("has/slash").is_err());
-        assert!(create("has.dot").is_err());
+        assert!(patina::mother::lake_runtime::validate_lake_name("").is_err());
+        assert!(patina::mother::lake_runtime::validate_lake_name("has spaces").is_err());
+        assert!(patina::mother::lake_runtime::validate_lake_name("has/slash").is_err());
+        assert!(patina::mother::lake_runtime::validate_lake_name("has.dot").is_err());
     }
 
     #[test]
     fn valid_lake_name_characters() {
-        assert!(validate_lake_name("good-name").is_ok());
-        assert!(validate_lake_name("my_lake").is_ok());
-        assert!(validate_lake_name("lake123").is_ok());
-        assert!(validate_lake_name("").is_err());
-        assert!(validate_lake_name("has spaces").is_err());
-        assert!(validate_lake_name("has.dot").is_err());
+        assert!(patina::mother::lake_runtime::validate_lake_name("good-name").is_ok());
+        assert!(patina::mother::lake_runtime::validate_lake_name("my_lake").is_ok());
+        assert!(patina::mother::lake_runtime::validate_lake_name("lake123").is_ok());
     }
 
     #[test]
