+++ src/mother/mod.rs
@@ -26,7 +26,9 @@
 //! ```
 
 pub mod broker;
+pub mod doctor_runtime;
 mod internal;
+pub mod lake_runtime;
 
 // Bridge module: state logic now lives in the mother crate.
 // Re-export so existing `crate::mother::state::*` paths continue to compile.
